Resolves issue #3515:
Our controllers did not invalidate cached values on parent value changed, so would not update when manually changing parent value.

To fix, I implemented parent value cache for Mat3 controllers, and invalidate other controllers when getting relative value

+++ src/MaxConversionFns.cpp
@@ -341,7 +341,7 @@ FabricCore::RTVal GetRTVal(const Mesh& param)
 template<>
 void SpliceToMaxValue(FabricCore::RTVal& rtVal, int& param)
 {
-	param = static_cast<float>(rtVal.getSInt32());
+	param = rtVal.getSInt32();
 }
 
 template<>
